+++ src/components/SearchBar/searchBarChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Updates styles on mobile breakpoints, adds span element to text for CSS targeting.",
+    ],
+  },
+  {
     date: "2024-06-20",
     version: "3.1.6",
     type: "Update",